+++ lua/core/bootstrap-plugin-manager.lua
