+++ lua/core/keymaps/async/tsserver.lua
@@ -1,0 +1,8 @@
+return function(buffer, cmds)
+  local nmap = require("core.utils").createNmap({
+    buffer = buffer,
+    desc = 'LSP: ',
+  })
+
+  nmap('<leader>oi', cmds.organize_imports, 'OrganizeImports')
+end
