+++ lua/core/keymaps/init.lua
@@ -1,0 +1,6 @@
+local import = require('core.utils').createImporter("core.keymaps")
+
+import('general')
+import('project-tree')
+import('file-search')
+import('diagnostics')
