+++ lua/core/keymaps/project-tree.lua
@@ -1,0 +1,4 @@
+-- Neotree mappings
+
+vim.keymap.set('n', '<leader><space>', ':Neotree toggle<cr>', { desc = 'File tree' })
+vim.keymap.set('n', '<leader>nf', ':Neotree filesystem reveal<cr>', { desc = 'Reveal in file tree' })
