+++ lua/core/setup/init.lua
@@ -1,0 +1,6 @@
+local import = require('core.utils').createImporter("core.setup")
+
+import('file-search')
+import('syntax-highlight')
+import('language-servers')
+import('completion-engine')
